dynamic_lighting
I added a couple test sprites so we'd have things to illuminate and illuminate with. I also implemented some dynamic lighting.
As for changing default light amounts, we should probably do that by using a large mix lighting node that covers the area we would want to give the default lighting to.